+++ src/firewallfabrik/__init__.py
@@ -10,4 +10,4 @@
 #
 # SPDX-License-Identifier: GPL-2.0-or-later
 
-__version__ = '1.9.173'
+__version__ = '1.9.174'